Port zdb
Slightly nonfunctional so far, it fails to find any pools.

It needs further inspection and porting, but for now just need the
source files to exist, for cherry-picking commits

+++ ZFSin/zfs/module/zfs/spa_config.c
@@ -154,7 +154,7 @@ out:
 static void
 spa_config_write(spa_config_dirent_t *dp, nvlist_t *nvl)
 {
-	size_t buflen;
+	uint32_t buflen;
 	char *buf;
 	vnode_t *vp;
 	int oflags = FWRITE | FTRUNC | FCREAT | FOFFMAX;
